IGAPP-11173: Move responsibilities
