+++ native/src/routes/Pois.tsx
@@ -1,15 +1,12 @@
 import { BottomSheetScrollViewMethods } from '@gorhom/bottom-sheet'
-import MapboxGL from '@react-native-mapbox-gl/maps'
-import React, { ReactElement, useCallback, useEffect, useRef, useState } from 'react'
+import React, { ReactElement, useCallback, useRef, useState } from 'react'
 import { useTranslation } from 'react-i18next'
-import { ScrollView, useWindowDimensions } from 'react-native'
+import { ScrollView } from 'react-native'
 import { useTheme } from 'styled-components'
 import styled from 'styled-components/native'
 
 import {
-  animationDuration,
   CityModel,
-  normalDetailZoom,
   embedInCollection,
   ErrorCode,
   fromError,
@@ -19,7 +16,6 @@ import {
   POIS_ROUTE,
   PoisRouteType,
   prepareFeatureLocations,
-  closerDetailZoom,
 } from 'api-client'
 
 import BottomActionsSheet from '../components/BottomActionsSheet'
@@ -44,7 +40,7 @@ const NoItemsMessage = styled.Text`
   text-align: center;
 `
 
-const midSnapPointPercentage = 0.35
+export const midSnapPointPercentage = 0.35
 const percentage = 100
 const BOTTOM_SHEET_SNAP_POINTS = [
   dimensions.bottomSheetHandler.height,
@@ -68,13 +64,11 @@ const Pois = ({ pois, language, cityModel, route, navigation }: PoisProps): Reac
   const [sheetSnapPointIndex, setSheetSnapPointIndex] = useState<number>(1)
   const [followUserLocation, setFollowUserLocation] = useState<boolean>(false)
   const [listScrollPosition, setListScrollPosition] = useState<number>(0)
-  const deviceHeight = useWindowDimensions().height
   const features = prepareFeatureLocations(pois, coordinates)
   const selectedFeature = slug ? features.find(it => it.properties.slug === slug) : null
   const poi = pois.find(it => it.slug === slug)
   const { t } = useTranslation('pois')
   const theme = useTheme()
-  const cameraRef = useRef<MapboxGL.Camera | null>(null)
   const scrollRef = useRef<BottomSheetScrollViewMethods>(null)
 
   const scrollTo = (position: number) => {
@@ -88,7 +82,7 @@ const Pois = ({ pois, language, cityModel, route, navigation }: PoisProps): Reac
     }, RESTORE_TIMEOUT)
   }
   const selectPoiFeature = (feature: PoiFeature | null) => {
-    if (feature && cameraRef.current) {
+    if (feature) {
       setFollowUserLocation(false)
       navigation.setParams({ slug: feature.properties.slug })
       scrollTo(0)
@@ -102,19 +96,6 @@ const Pois = ({ pois, language, cityModel, route, navigation }: PoisProps): Reac
     scrollTo(listScrollPosition)
   }, [listScrollPosition, navigation])
   useOnBackNavigation(slug ? deselectPoiFeature : undefined)
-
-  // Wait for followUserLocation change before moving the camera to avoid position lock
-  // https://github.com/rnmapbox/maps/issues/1079
-  useEffect(() => {
-    if (!followUserLocation && selectedFeature && cameraRef.current) {
-      cameraRef.current.setCamera({
-        centerCoordinate: selectedFeature.geometry.coordinates,
-        zoomLevel: selectedFeature.properties.closeToOtherPoi ? closerDetailZoom : normalDetailZoom,
-        animationDuration,
-        padding: { paddingBottom: deviceHeight * midSnapPointPercentage },
-      })
-    }
-  }, [deviceHeight, followUserLocation, selectedFeature])
 
   const renderPoiListItem = (poi: PoiFeature): ReactElement => (
     <PoiListItem
@@ -168,7 +149,6 @@ const Pois = ({ pois, language, cityModel, route, navigation }: PoisProps): Reac
   return (
     <ScrollView contentContainerStyle={{ flexGrow: 1 }}>
       <MapView
-        ref={cameraRef}
         selectPoiFeature={selectPoiFeature}
         boundingBox={cityModel.boundingBox}
         setSheetSnapPointIndex={setSheetSnapPointIndex}
